solved the problem!! derivative error

<에러 내용>
org.springframework.validation.BindException: org.springframework.validation.BeanPropertyBindingResult: 1 errors
Field error in object 'derivative' on field 'ordersid': rejected value [9]; codes [typeMismatch.derivative.ordersid,typeMismatch.ordersid,typeMismatch.com.olive.aio.orders.Orders,typeMismatch]; arguments [org.springframework.context.support.DefaultMessageSourceResolvable: codes [derivative.ordersid,ordersid]; arguments []; default message [ordersid]]; default message [Failed to convert property value of type 'java.lang.String' to required type 'com.olive.aio.orders.Orders' for property 'ordersid'; nested exception is org.springframework.core.convert.ConversionFailedException: Failed to convert from type [java.lang.String] to type [@javax.persistence.OneToOne @javax.persistence.JoinColumn com.olive.aio.orders.Orders] for value '9'; nested exception is org.springframework.dao.InvalidDataAccessApiUsageException: Provided id of the wrong type for class com.olive.aio.orders.Orders. Expected: class java.lang.Long, got class java.lang.Integer; nested exception is java.lang.IllegalArgumentException: Provided id of the wrong type for class com.olive.aio.orders.Orders. Expected: class java.lang.Long, got class java.lang.Integer]

에러내용은 앞이 아니라 뒤에 있다... String으로 넘어오는 파라메타값을 long으로 변환을 못하는게 아니라 long이 아 Integer로 변환이 됐다...
이유는 ... 내가 처음에 testorders 만들었을때 id값을 Integer로 작업하고 있어서 JparRpository를 long 타입이 아니라 Integer로 작업하고 있었는데
merge하고 기존에 쓰던 repository에서 entity이름만 변경해줌.....

public interface DerivativeOrdersRepository extends JpaRepository<Orders, Integer> 을
public interface DerivativeOrdersRepository extends JpaRepository<Orders, Long>
로 바꾸니까 잘된다.......ㅠㅠㅠㅠㅠㅠㅠㅠㅠㅠㅠㅠㅠㅠㅠㅠㅠ

+++ src/main/java/com/olive/aio/logisticsManage/derivative/DerivativeOrdersRepository.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Transactional(readOnly = true)
-public interface DerivativeOrdersRepository extends JpaRepository<Orders, Integer>{
+public interface DerivativeOrdersRepository extends JpaRepository<Orders, Long>{
 
     // 결제완료 발주리스트 조회
     List<Orders> findByOrdersstateOrderByOrdersidAsc(String state);
